Made analysis modular
Defined an 'analysisFunctions.py' module that has all the necessary functiions to do the bout analysis, to reduce clutter in the main analysis file.

The 'Bout Extraction and Analysis.py' file has also been rewritten a bit to store all the data from each fish into csv files. These can be generated for each day and for each genotype, however the experiment is organised. This allows people unfamiliar with Python to still be able to do the analysis using the data from the csv files.

The bout_indices have also been made into a separate csv file, if ever there needs to be any verification to be done at a later point to detect bugs. As far as the code is concerned, right now, the bout detection is working perfectly. I have verified it by plotting bout detection plots (see the Bout Extraction and Analysis.py). Each bout is shaded in green, and all bouts seem to be detected correctly.

'Plot Generation.py' has been made, but needs to be made modular, too. That will happen in a later update.